requirements.txt
pandas 
scikit-learn==1.1.3
numpy
flask
gunicorn
tensorflow-cpu